+++ src/const/message.ts
@@ -5,3 +5,5 @@ export const MESSAGE_CANCEL_FLAT = 'canceled';
 export const MESSAGE_THREAD_DIVIDER_ID = '__THREAD_DIVIDER__';
 
 export const MESSAGE_WELCOME_GUIDE_ID = 'welcome';
+
+export const THREAD_DRAFT_ID = '__THREAD_DRAFT_ID__';