+++ src/database/server/models/_template.ts
@@ -3,7 +3,7 @@ import { and, desc } from 'drizzle-orm/expressions';
 
 import { serverDB } from '@/database/server';
 
-import { NewSessionGroup, UserItem, sessionGroups } from '../schemas/lobechat';
+import { NewSessionGroup, SessionGroupItem, sessionGroups } from '../schemas/lobechat';
 
 export class TemplateModel {
   private userId: string;
@@ -44,7 +44,7 @@ export class TemplateModel {
     });
   };
 
-  async update(id: string, value: Partial<UserItem>) {
+  async update(id: string, value: Partial<SessionGroupItem>) {
     return serverDB
       .update(sessionGroups)
       .set({ ...value, updatedAt: new Date() })