+++ src/features/Conversation/hooks/useChatListActionsBar.tsx
@@ -1,9 +1,12 @@
 import { ActionIconGroupItems } from '@lobehub/ui/es/ActionIconGroup';
-import { Copy, Edit, ListRestart, RotateCcw, Trash } from 'lucide-react';
+import { Copy, Edit, ListRestart, RotateCcw, Split, Trash } from 'lucide-react';
 import { useMemo } from 'react';
 import { useTranslation } from 'react-i18next';
 
+import { isServerMode } from '@/const/version';
+
 interface ChatListActionsBar {
+  branching: ActionIconGroupItems;
   copy: ActionIconGroupItems;
   del: ActionIconGroupItems;
   delAndRegenerate: ActionIconGroupItems;
@@ -12,11 +15,21 @@ interface ChatListActionsBar {
   regenerate: ActionIconGroupItems;
 }
 
-export const useChatListActionsBar = (): ChatListActionsBar => {
+export const useChatListActionsBar = ({
+  hasThread,
+}: { hasThread?: boolean } = {}): ChatListActionsBar => {
   const { t } = useTranslation('common');
 
   return useMemo(
     () => ({
+      branching: {
+        disable: !isServerMode,
+        icon: Split,
+        key: 'branching',
+        label: isServerMode
+          ? t('branching', { defaultValue: 'Create Sub Topic' })
+          : t('branchingDisable'),
+      },
       copy: {
         icon: Copy,
         key: 'copy',
@@ -24,11 +37,13 @@ export const useChatListActionsBar = (): ChatListActionsBar => {
       },
       del: {
         danger: true,
+        disable: hasThread,
         icon: Trash,
         key: 'del',
-        label: t('delete', { defaultValue: 'Delete' }),
+        label: hasThread ? t('messageAction.deleteDisabledByThreads', { ns: 'chat' }) : t('delete'),
       },
       delAndRegenerate: {
+        disable: hasThread,
         icon: ListRestart,
         key: 'delAndRegenerate',
         label: t('messageAction.delAndRegenerate', {
@@ -50,6 +65,6 @@ export const useChatListActionsBar = (): ChatListActionsBar => {
         label: t('regenerate', { defaultValue: 'Regenerate' }),
       },
     }),
-    [],
+    [hasThread],
   );
 };